+++ gcpy/examples/bpch_to_nc/__init__.py
@@ -1,0 +1,7 @@
+"""
+GCPy import script
+"""
+
+from .bpch2nc import *
+from .bpch_tagco_prodloss_to_nc import *
+