+++ gcpy/examples/diagnostics/__init__.py
@@ -1,0 +1,6 @@
+"""
+GCPy import script
+"""
+
+from .compare_diags import *
+