+++ gcpy/examples/working_with_files/__init__.py
@@ -1,0 +1,7 @@
+"""
+GCPy import script
+"""
+from .add_blank_var_to_restart_file import *
+from .concatenate_files import *
+from .insert_field_into_restart_file import *
+from .regrid_restart_ll_to_cs import *